fix: close the review-cycle findings
An independent code review and a security review of this branch found three
Major issues and eleven smaller ones. Each was verified against the vendored
libjpeg sources or `std` before being acted on.

**A warning is not always a repair.** `is_accepted` refused any page carrying a
libjpeg warning outside the truncation set, and a page failure ends the run, so
one such page refused the whole archive with a message saying libjpeg had
repaired damage. Three of libjpeg's ten `JWRN_*` codes describe a
non-conforming header it interpreted rather than data it substituted, and
libjpeg says so itself: of `JWRN_NOT_SEQUENTIAL`, `jdhuff.c` reads "This ought
to be an error condition, but we make it a warning because there are some
baseline files out there with all zeroes in these bytes". Every code is now
classified as repair, benign, or truncation-shaped, an unlisted code is refused
so a future warning fails closed, and a unit test asserts the three lists are
disjoint and cover libjpeg's table.

**The truncation exception was wider than it looked.** Accepting a
truncation-shaped code set on a stream with no end-of-image marker is
satisfiable on purpose: corrupt an entropy byte at one of the offsets that
yields such a set, then strip the closing marker. Refusing
`JWRN_EXTRANEOUS_DATA` and requiring `JWRN_JPEG_EOF` narrows the measured
overlap from 266 of ~105,000 corruptions to 11. The residue is documented rather
than papered over: it separates accidental truncation from accidental
corruption, and the cost of that bound is a damaged page in an output archive,
not memory safety.

**The partial file followed links.** `Sink::create` checked only the final
output and opened `<output>.part` with `File::create`. The partial's name is
derived from the output's, so anyone able to write that directory could
pre-place it as a symbolic or hard link and have this process truncate the
target and then rename the link into place. It is now created with `create_new`,
and an existing partial is its own error.

Also fixed:

- Warnings are read through the fork's new `finish_with_warnings`.
  `jpeg_finish_decompress` reads on to the end-of-image marker and warns there
  about extraneous data, so a snapshot taken before finishing can be short.
- `Sink` gained a `Drop` that removes the partial unless the rename succeeded.
  `finish` consumes the sink, so its own failures — a full disk while the
  central directory is written — previously left a full-size stray file with no
  caller left to clean up.
- Each spawned stage catches its own panic. `thread::scope` re-panics for a
  spawned thread that panicked, which discarded the payload, skipped the error
  path, and bypassed cleanup; `RunError::StagePanicked` was unreachable.
- `Budget` computes in `u128`, as `PageImage::new` already did and for the same
  reason: two `u32` axes times three channels overflows `u64`.
- An entry's declared size no longer reserves up to 64 MiB. Up to `2 * jobs`
  such buffers are alive at once, so the hint is capped at a megabyte and
  `read_to_end` grows from there.
- Two stored names that collide once renamed to the encoder's extension are
  reported as a collision instead of as `zip`'s "Duplicate filename".
- An archive with no pages is refused instead of installing an empty archive
  that would report success and then make the next run fail with "already
  exists".
- A traversing, absolute, drive-lettered, or NUL-bearing entry name is refused
  rather than carried into the output archive for someone else's extractor.
- The `MAX_IMAGE_BYTES` note no longer understates the peak: it bounds one
  buffer, a worker holds two, and libjpeg's own working set follows the source
  geometry.
- `run`'s documentation says which failure is reported when several pages fail.
- `the_output_holds_only_image_entries` reads the output with `zip::ZipArchive`,
  because reading it back through this crate's own reader skips non-image
  entries and so could not observe one being written.

+++ src/page/budget.rs
@@ -26,7 +26,13 @@ const MAX_SOURCE_PIXELS: u64 = 100_000_000;
 ///
 /// Chosen, not measured. A 1280-wide page is about 7 MB decoded; a 600dpi spread about
 /// 210 MB. 256 MiB admits that and refuses the 12.87 GB a `65500x65500` RGB header asks
-/// for. Peak memory is this times the worker count, so it is not a free number to raise.
+/// for.
+///
+/// Raising it is not free, and it is not the only term either: this bounds *one buffer*, a
+/// worker holds a decoded page and a resize destination at once, and libjpeg's own working
+/// set follows the source geometry rather than this — a progressive source at the pixel
+/// ceiling holds coefficient arrays of roughly 600 MB inside libjpeg alone, with no backing
+/// store, because mozjpeg links `jmemnobs`. Multiply the lot by the worker count.
 const MAX_IMAGE_BYTES: u64 = 256 << 20;
 
 /// What one page is allowed to cost.
@@ -60,16 +66,13 @@ impl Budget {
 
     /// Rejects a page whose header declares more pixels than the limit allows.
     ///
-    /// Widened to `u64`: two `u32` axes overflow `u32`, and a release build has overflow
-    /// checks off, so a wrapped product would accept the very header this refuses.
-    ///
     /// # Errors
     ///
     /// [`PageErrorKind::TooLarge`], naming the quantity, the value, and the limit.
     pub fn allow_source(&self, width: u32, height: u32) -> Result<(), PageErrorKind> {
         Self::check(
             "source pixels",
-            u64::from(width) * u64::from(height),
+            u128::from(width) * u128::from(height),
             self.max_source_pixels,
         )
     }
@@ -87,13 +90,20 @@ impl Budget {
     ) -> Result<(), PageErrorKind> {
         Self::check(
             "image bytes",
-            u64::from(width) * u64::from(height) * u64::from(channels.count()),
+            u128::from(width) * u128::from(height) * u128::from(channels.count()),
             self.max_image_bytes,
         )
     }
 
-    fn check(quantity: &'static str, actual: u64, limit: u64) -> Result<(), PageErrorKind> {
-        if actual > limit {
+    /// Widened to `u128`, as [`PageImage::new`] is and for the same reason: two `u32` axes
+    /// times three channels tops out just under `3 * u64::MAX`, a release build has overflow
+    /// checks off, and a wrapped product would compare below the limit and pass. Not
+    /// reachable through the binary, whose width is capped and whose JPEG axes are 16-bit,
+    /// but this check is what stands between a library caller and an infallible `vec![0; …]`.
+    ///
+    /// [`PageImage::new`]: super::PageImage::new
+    fn check(quantity: &'static str, actual: u128, limit: u64) -> Result<(), PageErrorKind> {
+        if actual > u128::from(limit) {
             return Err(PageErrorKind::TooLarge {
                 quantity,
                 actual,
